+++ Develop/package.json
@@ -3,17 +3,28 @@
   "version": "1.0.0",
   "main": "app.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "start": "node app.js"
   },
   "keywords": [],
-  "author": "",
-  "license": "ISC",
+  "author": "Anthony Gillespie",
+  "license": "MIT",
   "dependencies": {
     "inquirer": "^12.0.0",
     "pg": "^8.13.0"
   },
   "devDependencies": {
+    "@types/node": "^22.7.8",
+    "@types/pg": "^8.11.10",
     "typescript": "^5.6.3"
   },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/UnknownKonflikt/Manager_Command.git"
+  },
+  "bugs": {
+    "url": "https://github.com/UnknownKonflikt/Manager_Command/issues"
+  },
+  "homepage": "https://github.com/UnknownKonflikt/Manager_Command#readme",
   "description": ""
 }
